requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0
shap>=0.42.0
streamlit>=1.25.0
pydantic>=2.0.0
pytest>=7.4.0
black>=23.0.0
ruff>=0.0.280
pre-commit>=3.3.0
tensorboard>=2.13.0
optuna>=3.2.0
imbalanced-learn>=0.11.0
calibration-toolbox>=0.1.0
lifelines>=0.27.0
pytorch-tabnet>=4.0
pytorch-tabular>=1.0.0
